backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
marshmallow==3.20.1
PyJWT==2.8.0
bcrypt==4.0.1
pytest==7.4.2
pytest-flask==1.2.0 